refactor(session): extract the one-shot session actions to sessionActions.ts
downloadNec / loadMeasured / copyParams run entirely on a user gesture and
own no React state, so they leave DesignSession as plain async functions
taking their setters and buildRequest as arguments. Bodies are verbatim;
the three JSX call sites gain arrow wrappers to supply the arguments.
